feat: Add ApiController
* neuron-ui talks to remote/api instead of remote/app
* AppController is used as instance, not static functions object
* Detach menu from app controller
* Remove controller and UI action dependency from services
