+++ packages/neuron-wallet/src/controllers/wallets.ts
@@ -1,5 +1,5 @@
 import fs from 'fs'
-import { dialog, SaveDialogReturnValue } from 'electron'
+import { dialog, SaveDialogReturnValue, BrowserWindow } from 'electron'
 import WalletsService, { Wallet, WalletProperties, FileKeystoreWallet } from 'services/wallets'
 import Keystore from 'models/keys/keystore'
 import Keychain from 'models/keys/keychain'
@@ -21,7 +21,6 @@ import i18n from 'utils/i18n'
 import AddressService from 'services/addresses'
 import WalletCreatedSubject from 'models/subjects/wallet-created-subject'
 import logger from 'utils/logger'
-import AppController from './app'
 
 /**
  * @class WalletsController
@@ -286,7 +285,7 @@ export default class WalletsController {
 
     const keystore = wallet.loadKeystore()
     return new Promise(resolve => {
-      dialog.showSaveDialog(AppController.mainWindow!, { title: i18n.t('messages.save-keystore'), defaultPath: wallet.name + '.json' }).then(
+      dialog.showSaveDialog(BrowserWindow.getFocusedWindow()!, { title: i18n.t('messages.save-keystore'), defaultPath: wallet.name + '.json' }).then(
         (returnValue: SaveDialogReturnValue) => {
           if (returnValue.filePath) {
             fs.writeFileSync(returnValue.filePath, JSON.stringify(keystore))
